+++ internal/provider/provider.go
@@ -401,7 +401,6 @@ func (p *CcProvider) DataSources(ctx context.Context) []func() datasource.DataSo
 		NewTransitNetworkDataSource,
 		NewUpdateAuthenticationProfileDataSource,
 		NewUserDataSource,
-		NewVirtualNetworkToFabricSiteDataSource,
 		NewWirelessEnterpriseSSIDDataSource,
 		NewWirelessInterfaceDataSource,
 		NewWirelessProfileDataSource,