+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.2.15 (unreleased)
 
+- BREAKING CHANGE: Modified `catalystcenter_virtual_network_to_fabric_site` to use the `/dna/intent/api/v1/sda/layer3VirtualNetworks` API endpoint instead of `/dna/intent/api/v1/business/sda/virtual-network` and remove data_source and import
 - Add a debug flag to enable starting a provider in debug mode
 - Fix idempotency issue with `catalystcenter_telemetry_settings`